+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### master
+- if the pane is too narrow do not show sidebar
 
 ### v0.2.0, Sep 03, 2014
 - automatic sidebar opening for another sidebar type